% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPhylo_speciesRaster.R
\name{addPhylo_speciesRaster}
\alias{addPhylo_speciesRaster}
\title{addPhylo_speciesRaster}
\usage{
addPhylo_speciesRaster(x, tree, replace = FALSE)
}
\arguments{
\item{x}{object of class \code{speciesRaster}}

\item{tree}{a phylogeny of class \code{phylo}}

\item{replace}{boolean; if a tree is already a part of \code{x},
should it be replaced?}
}
\value{
object of class \code{speciesRaster}, with a \code{phylo}
object as the list element named \code{phylo}.
}
\description{
Add a phylogeny to speciesRaster object.
}
\details{
If any species in the phylogeny are not found in the speciesRaster
geographical data, then those species will be dropped from the phylogeny, and
a warning will be issued.
}
\examples{
tamiasSpRas
tamiasTree

addPhylo_speciesRaster(tamiasSpRas, tamiasTree)

}
\author{
Pascal Title
}
