% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTraits_speciesRaster.R
\name{addTraits_speciesRaster}
\alias{addTraits_speciesRaster}
\title{addTraits_speciesRaster}
\usage{
addTraits_speciesRaster(x, data, replace = FALSE)
}
\arguments{
\item{x}{object of class \code{speciesRaster}}

\item{data}{named numeric vector, matrix or dataframe with rownames corresponding to species in \code{x}
or pairwise matrix with row and column names corresponding to species in \code{x}. If pairwise matrix,
the upper triangle of the matrix will be used for calculations.}

\item{replace}{boolean; if data is already a part of \code{x},
should it be replaced?}
}
\value{
object of class \code{speciesRaster}, with trait data
as the list element named \code{data}.
}
\description{
Add univariate or multivariate trait data to a speciesRaster object.
}
\details{
If any species in \code{data} are not found in the speciesRaster
geographical data, then those species will be dropped from \code{data}, and
a warning will be issued.
}
\examples{
tamiasSpRas
tamiasTraits

addTraits_speciesRaster(tamiasSpRas, tamiasTraits)

}
\author{
Pascal Title
}
