% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDiversity2.R
\name{betaDiversity2_speciesRaster}
\alias{betaDiversity2_speciesRaster}
\title{Map turnover in species communities v2}
\usage{
betaDiversity2_speciesRaster(x, radius, metric, component = "full")
}
\arguments{
\item{x}{object of class \code{speciesRaster}.}

\item{radius}{Radius of the moving window in map units.}

\item{metric}{choice of metric, see details.}

\item{component}{which component of beta diversity to use, can be \code{"turnover"}, 
\code{"nestedness"} or \code{"full"}}
}
\value{
Returns a raster with mean community dissimilarity for each cell.
}
\description{
Mean community dissimilarity in terms of species composition, phylogeny or traits is 
	calculated for each cell within a circular moving window of neighboring cells.
}
\details{
For each cell, mean dissimilarity is calculated from the focal cell to each of its neighbors.

	The following metrics are available. All metrics are based on Sorensen dissimilarity and range 
	from 0 to 1:
	\itemize{
		\item{taxonomic}
		\item{phylogenetic}
		\item{trait}
	}
	For each metric, the following components can be specified. These components are additive,
	such that the full metric = turnover + nestedness. 
	\itemize{
		\item{turnover}: species turnover without the influence of richness differences
		\item{nestedness}: species turnover due to differences in richness
		\item{full}: the combined turnover due to both differences in richness and pure turnover
	}

Trait turnover \code{trait} measures the turnover in community trait data. This metric is 
identical to phylogenetic turnover, but where the phylogeny is replaced with a neighbor-joining tree 
that is generated from the trait distance matrix.

The following are currently not implemented, but might be in the future:
Range-weighted turnover \code{RWTurnover} measures turnover but where taxa are weighted
	according to the inverse of their range size.

Phylogenetic range-weighted turnover \code{phyloRWTurnover} measures turnover in phylogenetic diversity
	where phylogenetic branches are weighted by the inverse of their geographic distribution.
}
\examples{
\donttest{
library(raster)
tamiasSpRas

tamiasSpRas <- addPhylo_speciesRaster(tamiasSpRas, tamiasTree)
tamiasSpRas <- addTraits_speciesRaster(tamiasSpRas, tamiasTraits)

# taxonomic turnover
beta_taxonomic_turnover <- betaDiversity2_speciesRaster(tamiasSpRas, radius = 70000, 
	metric = 'taxonomic', component = 'turnover')
beta_taxonomic_nestedness <- betaDiversity2_speciesRaster(tamiasSpRas, radius = 70000, 
	metric = 'taxonomic', component = 'nestedness')
beta_taxonomic_full <- betaDiversity2_speciesRaster(tamiasSpRas, radius = 70000, 
	metric = 'taxonomic', component = 'full')

# phylogenetic turnover
beta_phylo_turnover <- betaDiversity2_speciesRaster(tamiasSpRas, radius = 70000, 
	metric = 'phylogenetic', component = 'turnover')
beta_phylo_nestedness <- betaDiversity2_speciesRaster(tamiasSpRas, radius = 70000, 
	metric = 'phylogenetic', component = 'nestedness')
beta_phylo_full <- betaDiversity2_speciesRaster(tamiasSpRas, radius = 70000, 
	metric = 'phylogenetic', component = 'full')

# trait turnover
beta_trait_turnover <- betaDiversity2_speciesRaster(tamiasSpRas, radius = 70000, 
	metric = 'trait', component = 'turnover')
beta_trait_nestedness <- betaDiversity2_speciesRaster(tamiasSpRas, radius = 70000, 
	metric = 'trait', component = 'nestedness')
beta_trait_full <- betaDiversity2_speciesRaster(tamiasSpRas, radius = 70000, 
	metric = 'trait', component = 'full')


colramp <- colorRampPalette(c('blue','yellow','red'))
par(mfrow=c(1,3))
plot(beta_taxonomic_full, col = colramp(100), zlim = c(0,1))
plot(beta_phylo_full, col = colramp(100), zlim = c(0,1))
plot(beta_trait_full, col = colramp(100), zlim = c(0,1))

}
}
\references{
Baselga, A. The relationship between species replacement, dissimilarity derived from nestedness, 
and nestedness. Global Ecology and Biogeography 21 (2012): 1223–1232.

Laffan, SW, et al. Range-weighted metrics of species and phylogenetic turnover can better 
resolve biogeographic transition zones. Methods in Ecology and Evolution 7 (2016): 580-588.

Leprieur, F, Albouy, C, De Bortoli, J, Cowman, PF, Bellwood, DR & Mouillot, D. Quantifying 
Phylogenetic Beta Diversity: Distinguishing between "True" Turnover of Lineages and Phylogenetic 
Diversity Gradients. PLoS ONE 7 (2012): e42760–12.

Rosauer, D, Laffan, SW, Crisp, MD, Donnellan, SC, Cook, LG. Phylogenetic endemism: a new approach 
for identifying geographical concentrations of evolutionary history. Molecular Ecology 
18 (2009): 4061-4072.
}
\author{
Pascal Title
}
