% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarsen_speciesRaster.R
\name{coarsen_speciesRaster}
\alias{coarsen_speciesRaster}
\title{Coarsen a speciesRaster}
\usage{
coarsen_speciesRaster(x, fact)
}
\arguments{
\item{x}{object of class \code{speciesRaster}}

\item{fact}{the aggregation factor as the number of cells in the horizontal
and vertical directions.}
}
\value{
\code{speciesRaster} with a coarser resolution.
}
\description{
Return a speciesRaster object with a coarser resolution.
}
\details{
Species will be included in the coarser cell if they are found in
	>= 50 percent of the original-resolution cells that are being aggregated. 
	This should be thought of as more of a convenience function for experimentation.
	Creating a coarser resolution speciesRaster from species ranges with 
	\code{\link{createSpeciesRaster}} is the more appropriate approach.
}
\examples{
tamiasSpRas

# coarsen such that groups of 4 cells are aggregated to 1 cell
# (2 cells in horizontal and 2 cells in vertical directions)
coarsen_speciesRaster(tamiasSpRas, fact = 2)

}
\author{
Pascal Title
}
