% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropFromSpeciesRaster.R
\name{dropFromSpeciesRaster}
\alias{dropFromSpeciesRaster}
\title{Drop species from speciesRaster}
\usage{
dropFromSpeciesRaster(x, sp)
}
\arguments{
\item{x}{object of class \code{speciesRaster}}

\item{sp}{a character vector of species names to be dropped.}
}
\value{
new \code{speciesRaster} object.
}
\description{
Removes particular species from a speciesRaster object.
}
\details{
If species in \code{sp} are not in \code{x}, they will be ignored.
}
\examples{
tamiasSpRas

newSpRas <- dropFromSpeciesRaster(tamiasSpRas, sp = c('Tamias_alpinus', 'Tamias_bulleri'))

setdiff(tamiasSpRas[['geogSpecies']], newSpRas[['geogSpecies']])


}
\author{
Pascal Title
}
