% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_speciesRaster.R
\name{getData_speciesRaster}
\alias{getData_speciesRaster}
\title{Get data from speciesRaster}
\usage{
getData_speciesRaster(x)
}
\arguments{
\item{x}{object of class \code{speciesRaster}}
}
\value{
The object found in the \code{data} location of the speciesRaster.
	Could be either a vector or a matrix.
}
\description{
Return the data item from a speciesRaster object.
}
\examples{
x <- addTraits_speciesRaster(tamiasSpRas, tamiasTraits)
getData_speciesRaster(x)

}
\author{
Pascal Title
}
