% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMultiMapRamp.R
\name{getMultiMapRamp}
\alias{getMultiMapRamp}
\title{Extract min and max for multiple speciesRasters}
\usage{
getMultiMapRamp(x)
}
\arguments{
\item{x}{list of objects of class \code{speciesRaster} or \code{RasterLayer}.}
}
\value{
a numeric vector of length 2: overall min and max value.
}
\description{
Extracts the range of values across a list of speciesRaster
	objects for use in plotting
}
\details{
If the user would like to plot multiple speciesRaster objects
with a standardized color ramp, then the returned values from this function
can be supplied to \code{\link{plot.speciesRaster}}. Also works with RasterLayer
objects.
}
\author{
Pascal Title
}
