% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhylo_speciesRaster.R
\name{getPhylo_speciesRaster}
\alias{getPhylo_speciesRaster}
\title{Get phylogeny from speciesRaster}
\usage{
getPhylo_speciesRaster(x)
}
\arguments{
\item{x}{object of class \code{speciesRaster}}
}
\value{
The phylogeny in the \code{phylo} location of the speciesRaster.
}
\description{
Return the phylo item from a speciesRaster object.
}
\examples{
x <- addPhylo_speciesRaster(tamiasSpRas, tamiasTree)
getPhylo_speciesRaster(x)

}
\author{
Pascal Title
}
