% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.speciesRaster.R
\name{identify.speciesRaster}
\alias{identify.speciesRaster}
\title{Identify species in speciesRaster}
\usage{
\method{identify}{speciesRaster}(x, ..., returnCell = FALSE)
}
\arguments{
\item{x}{object of class \code{speciesRaster}}

\item{...}{further arguments passed to \code{\link{identify}}}

\item{returnCell}{boolean; if FALSE, then species names are returned,
if TRUE, then cell numbers are returned.}
}
\value{
A vector of species names or cell numbers.
}
\description{
Returns the species found in a cell that is clicked on
in a plotted speciesRaster object.
}
\details{
This is a wrapper function for the \code{identify} function
in base graphics. This is primarily intended as a useful function for 
data exploration and spot-checking.
}
\author{
Pascal Title
}
