% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuildSpeciesRaster.R
\name{rebuildSpeciesRaster}
\alias{rebuildSpeciesRaster}
\title{Rebuild speciesRaster}
\usage{
rebuildSpeciesRaster(x, spCellList, rasterTemplate = NULL)
}
\arguments{
\item{x}{object of class \code{speciesRaster}}

\item{spCellList}{list in which each element is a character vector of species names.}

\item{rasterTemplate}{a separate raster can be provided if it has different 
characteristics than the one in \code{x}.}
}
\value{
new \code{speciesRaster} object.
}
\description{
Given a list of species in each cell, all components of the
	speciesRaster object are reindexed and regenerated.
}
\details{
This function is used internally by \code{\link{dropFromSpeciesRaster}} and 
	by \code{\link{coarsen_speciesRaster}}.
}
\author{
Pascal Title
}
