% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSpCellIndex.R
\name{singleSpCellIndex}
\alias{singleSpCellIndex}
\title{Identify single-species cells}
\usage{
singleSpCellIndex(x)
}
\arguments{
\item{x}{object of class \code{speciesRaster}}
}
\value{
numeric vector of raster cell numbers.
}
\description{
Given a speciesRaster object, return the raster cell indices
	of those cells that have just one species.
}
\details{
This function can be useful when further analyzing speciesRaster
	objects generated by \code{\link{cellMetrics_speciesRaster}}, as it might
	make sense to exclude these single-species cells in further analyses.
}
\examples{
singleSpCellIndex(tamiasSpRas)

}
\author{
Pascal Title
}
