% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{speciesRaster-example}
\alias{speciesRaster-example}
\alias{tamiasPolyList}
\alias{tamiasTree}
\alias{tamiasTraits}
\alias{tamiasSpRas}
\title{Chipmunk example dataset}
\description{
Range polygons, phylogeny, traits and speciesRaster of chipmunks
for illustration purposes.
}
