% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesRaster-package.R
\docType{package}
\name{speciesRaster}
\alias{speciesRaster}
\title{speciesRaster}
\description{
An R package that facilitates the aggregation of species'
geographic ranges from vector or raster spatial data, and that enables
the calculation of various morphological and phylogenetic community metics
across geography.
}
\author{
Pascal Title, Miriam Zelditch
}
