% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesRasterToPhyloComm.R
\name{speciesRasterToPhyloComm}
\alias{speciesRasterToPhyloComm}
\title{Convert speciesRaster to community matrix}
\usage{
speciesRasterToPhyloComm(x, sites)
}
\arguments{
\item{x}{object of class \code{speciesRaster}}

\item{sites}{locations of sites, see details.}
}
\value{
community matrix, with sites as rows and species as columns
}
\description{
Given specific sites, convert speciesRaster to 
	phylocomm matrix, with sites as rows, and species as columns
}
\details{
If sites are site coordinates, 
	then dataframe or matrix with two columns; 
	if sites are cell indices, then numeric vector;
	if \code{sites = 'all'}, then all cells will be returned as sites.
}
\examples{
library(raster)
tamiasSpRas

# from cell indices
cells <- c(22440, 41446, 39283, 37697)
speciesRasterToPhyloComm(tamiasSpRas, cells)

# from coordinates
xy <- xyFromCell(tamiasSpRas[[1]], cells)
speciesRasterToPhyloComm(tamiasSpRas, xy)

}
\author{
Pascal Title
}
