% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.speciesRaster.R
\name{summary.speciesRaster}
\alias{summary.speciesRaster}
\title{speciesRaster summary}
\usage{
\method{summary}{speciesRaster}(object)
}
\arguments{
\item{object}{object of class \code{speciesRaster}}
}
\value{
A list containing the summary information is returned invisibly.
}
\description{
Generates a summary of a speciesRaster object.
}
\details{
Summary information includes
}
\examples{
summary(tamiasSpRas)
attr <- summary(tamiasSpRas)
attr

}
\author{
Pascal Title
}
