% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableFromSpeciesRaster.R
\name{tableFromSpeciesRaster}
\alias{tableFromSpeciesRaster}
\title{Data table from speciesRaster}
\usage{
tableFromSpeciesRaster(..., n = NULL, dropSingleSpCells = TRUE)
}
\arguments{
\item{...}{objects of class \code{speciesRaster}, \code{RasterLayer} or \code{RasterStack}.}

\item{n}{number of cells to randomly subsample, no subsampling if \code{NULL}}

\item{dropSingleSpCells}{logical; should cells with single species be excluded?}
}
\value{
data.frame with input variables, as well as \code{"long"} and \code{"lat"}.
}
\description{
Given one or several speciesRaster objects, create a table of 
	values and associated coordinate data.
}
\details{
A set of cells are identified in the speciesRaster objects. If \code{n=NULL},
	then all cells are used, otherwise cells are randomly subsampled. Values at those
	cells are then returned. This table construction can be particularly useful for 
	subsequent statistical analyses.

	Only cells with data in all inputs are returned. If n is greater than the number
	of cells with data, then fewer than n cells will be returned.
}
\examples{

tamiasSpRas
tamiasSpRas <- addPhylo_speciesRaster(tamiasSpRas, tamiasTree)
tamiasSpRas <- addTraits_speciesRaster(tamiasSpRas, tamiasTraits)
morphoDisp <- cellMetrics_speciesRaster(tamiasSpRas, metric='disparity')
meanPat <- cellMetrics_speciesRaster(tamiasSpRas, metric='meanPatristic')

tableFromSpeciesRaster(tamiasSpRas, morphoDisp, meanPat, n = 100, dropSingleSpCells = TRUE)

}
\author{
Pascal Title
}
