% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSpeciesRaster.R
\name{writeSpeciesRaster}
\alias{writeSpeciesRaster}
\title{Save speciesRaster object}
\usage{
writeSpeciesRaster(x, filename)
}
\arguments{
\item{x}{object of class \code{speciesRaster}}

\item{filename}{filename with no extension}
}
\value{
Nothing is returned, but object is written to disk.
}
\description{
Write a speciesRaster object to disk.
}
\details{
This function ensures that the raster data is stored in memory
and then writes a .rds file. This file can be read back in with \code{\link{readRDS}}.
}
\examples{
\dontrun{
#save
save(tamiasSpRas, '~/tamiasSpeciesRaster')

# read back in
tamiasSpRas <- readRDS('~/tamiasSpeciesRaster.rds')
}
}
\author{
Pascal Title
}
