% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDiversity.R
\name{betaDiversity}
\alias{betaDiversity}
\title{Map turnover in beta diversity}
\usage{
betaDiversity(x, radius = 50000, metric = "betaSOR")
}
\arguments{
\item{x}{object of class \code{speciesRaster}.}

\item{radius}{Radius length in map units to define the moving window. 
This will be converted by rounding up to a number of neighboring 
cells in all directions.}

\item{metric}{choice of metric, see details.}
}
\value{
Returns a new \code{speciesRaster} object, with mean community dissimilarity for each cell.
}
\description{
Mean community dissimilarity is calculated for each cell within a moving window of neighboring cells.
}
\details{
Should put more information about the specific metric here, 
eventually implement a few complimentary ones.
Currently: Multi-site metrics implemented for: Sorensen dissimilarity (betaSOR), Simpson
dissimlarity / turnover component of Sorensen dissimilarity (betaSIM), nestedness-resultant
component of Sorensen dissimilarity (betaSNE), jaccard dissimilarity (betaJAC), turnover component
of jaccard dissimilarity (betaJTU), nestedness-resultant component of Jaccard dissimilarity
(betaJNE).
}
\examples{
tamiasSpRas

# using a moving window of roughly 50km radius
beta <- betaDiversity(tamiasSpRas, radius = 50000, metric = 'betaSOR')
plot(beta)

}
\author{
Pascal Title
}
