% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWTurnover.R
\name{betaDiversity_speciesRaster}
\alias{betaDiversity_speciesRaster}
\title{Map turnover in beta diversity}
\usage{
betaDiversity_speciesRaster(x, radius = 3, metric, verbose = FALSE)
}
\arguments{
\item{x}{object of class \code{speciesRaster}.}

\item{radius}{Radius in terms of number of cells in all directions from the focal cell 
to define the moving window.}

\item{metric}{choice of metric, see details.}

\item{verbose}{Primarily intended for debugging, print progress to the console.}
}
\value{
Returns a new \code{speciesRaster} object, with mean community dissimilarity for each cell.
}
\description{
Mean community dissimilarity is calculated for each cell within a moving window of neighboring cells.
}
\details{
Sorenson's Beta diversity \code{Sorenson} is a purely taxonomic measure of turnover.

Range-weighted turnover \code{RWTurnover} measures turnover but where taxa are weighted
	according to the inverse of their range size.

Phylogenetic range-weighted turnover \code{phyloRWTurnover} measures turnover in phylogenetic diversity
	where phylogenetic branches are weighted by the inverse of their geographic distribution.


The radius value defines the number of cells counted out from the focal cell in the horizontal and vertical
	direction, creating a square window of \code{side = radius * 2 + 1}. For each cell, all calculations
	between the focal cell and a moving window cell are run, and the results are averaged and assigned 
	to that focal cell.
}
\examples{
library(raster)
tamiasSpRas

beta_taxonomic <- betaDiversity_speciesRaster(tamiasSpRas, radius = 4, metric = 'Sorenson')

beta_RW <- betaDiversity_speciesRaster(tamiasSpRas, radius = 4, metric = 'RWTurnover')

tamiasSpRas <- addPhylo_speciesRaster(tamiasSpRas, tamiasTree)
beta_phyloRW <- betaDiversity_speciesRaster(tamiasSpRas, radius = 4, metric = 'phyloRWTurnover')

colramp <- colorRampPalette(c('blue','yellow','red'))
par(mfrow=c(1,3))
plot(beta_taxonomic, col = colramp(100))
plot(beta_RW, col = colramp(100))
plot(beta_phyloRW, col = colramp(100))


}
\references{
Laffan, SW, et al. Range-weighted metrics of species and phylogenetic turnover can better 
resolve biogeographic transition zones. Methods in Ecology and Evolution 7 (2016): 580-588.

Rosauer, D, Laffan, SW, Crisp, MD, Donnellan, SC, Cook, LG. Phylogenetic endemism: a new approach 
for identifying geographical concentrations of evolutionary history. Molecular Ecology 
18 (2009): 4061-4072.
}
\author{
Pascal Title
}
