% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellMetrics_speciesRaster.R
\name{cellMetrics_speciesRaster}
\alias{cellMetrics_speciesRaster}
\title{Cell Metrics}
\usage{
cellMetrics_speciesRaster(x, metric, var = NULL, nreps = 20,
  verbose = FALSE)
}
\arguments{
\item{x}{object of class \code{speciesRaster}}

\item{metric}{name of metric to use, see Details.}

\item{var}{If a univariate morphological metric is specified, and the 
data in \code{x} are multivariate, which trait should be used?
This can also specify which subset of columns a multivariate metric should be applied to.}

\item{nreps}{Number of repetitions for Foote metric distribution.}

\item{verbose}{Intended primarily for debugging, prints progress to the console}
}
\value{
object of class \code{speciesRaster} where the raster represents
	calculations of the metric at every cell.
}
\description{
Calculate various morphological and phylogenetic community
metrics for every cell in a \code{speciesRaster} object.
}
\details{
Univariate morphological metrics
	\itemize{
		\item{mean}
		\item{median}
		\item{range}
		\item{mean_NN_dist:} {mean nearest neighbor distance}
		\item{min_NN_dist:} {minimum nearest neighbor distance}
		\item{variance}
		\item{arithmeticWeightedMean} (see below)
		\item{geometricWeightedMean} (see below)
		\item{phylosignal:} {Blomberg's K for phylogenetic signal}
	}
	Multivariate morphological metrics
	\itemize{
		\item{disparity} 
		\item{range}
		\item{rangePCA}
		\item{mean_NN_dist:} {mean nearest neighbor distance}
		\item{min_NN_dist:} {minimum nearest neighbor distance}
		\item{phylosignal:} {Blomberg's K for phylogenetic signal, as implemented in geomorph::physignal}
	}
	Phylogenetic metrics
	\itemize{
		\item{pd:} {Faith's phylogenetic diversity, including the root}
		\item{meanPatristic}
		\item{meanPatristicNN:} {mean nearest neighbor in patristic distance}
		\item{minPatristicNN:} {minimum nearest neighbor in patristic distance}
		\item{phyloDisparity:} {sum of squared deviations in patristic distance}
		\item{PSV:} {Phylogenetic Species Variability}
	}
	Range-weighted metrics
	\itemize{
		\item{weightedEndemism}
		\item{correctedWeightedEndemism:} {Weighted endemism standardized by species richness}
		\item{phyloWeightedEndemism:}
	}

	If data slot contains a pairwise matrix, var is ignored.
	Weighted mean options are available where, for each cell, a weighting scheme 
	(inverse of species range sizes) is applied such that small-ranged species are 
	up-weighted, and broadly distributed species are down-weighted. 
	This can be a useful way to lessen the influence of broadly distributed species 
	in the geographic mapping of trait data.
}
\examples{
tamiasSpRas <- addPhylo_speciesRaster(tamiasSpRas, tamiasTree)
tamiasSpRas <- addTraits_speciesRaster(tamiasSpRas, tamiasTraits)

# univariate morphological example
x <- cellMetrics_speciesRaster(tamiasSpRas, metric='mean', var='V2')
plot(x) 

# multivariate morphological
x <- cellMetrics_speciesRaster(tamiasSpRas, metric='disparity')
plot(x)

# phylogenetic metrics
x <- cellMetrics_speciesRaster(tamiasSpRas, metric='meanPatristic')
plot(x)

}
