% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellMetrics_speciesRaster.R
\name{cellMetrics_speciesRaster}
\alias{cellMetrics_speciesRaster}
\title{Cell Metrics}
\usage{
cellMetrics_speciesRaster(x, metric, var = NULL, nreps = 20,
  verbose = FALSE)
}
\arguments{
\item{x}{object of class \code{speciesRaster}}

\item{metric}{name of metric to use, see Details.}

\item{var}{If a univariate morphological metric is specified, and the 
data in \code{x} are multivariate, which trait should be used?}

\item{nreps}{Number of repetitions for Foote metric distribution.}

\item{verbose}{Intended primarily for debugging, prints progress to the console}
}
\value{
object of class \code{speciesRaster} where the raster represents
	calculations of the metric at every cell.
}
\description{
Calculate various morphological and phylogenetic community
metrics for every cell in a \code{speciesRaster} object.
}
\details{
Univariate morphological metrics
	\itemize{
		\item{mean}
		\item{median}
		\item{range}
		\item{NN_dist:} {mean nearest neighbor distance}
		\item{variance}
	}
	Multivariate morphological metrics
	\itemize{
		\item{disparity} 
		\item{range}
		\item{rangePCA}
		\item{NN_dist:} {mean nearest neighbor distance}
	}
	Phylogenetic metrics
	\itemize{
		\item{meanPatristic}
		\item{patristicNN:} {mean nearest neighbor in patristic distance}
		\item{phyloDisparity:} {sum of squared deviations in patristic distance}
	}
}
\examples{
tamiasSpRas <- addPhylo_speciesRaster(tamiasSpRas, tamiasTree)
tamiasSpRas <- addTraits_speciesRaster(tamiasSpRas, tamiasTraits)

# univariate morphological example
x <- cellMetrics_speciesRaster(tamiasSpRas, metric='mean', var='V2')
plot(x) 

# multivariate morphological
x <- cellMetrics_speciesRaster(tamiasSpRas, metric='disparity')
plot(x)

# phylogenetic metrics
x <- cellMetrics_speciesRaster(tamiasSpRas, metric='meanPatristic')
plot(x)

}
