% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpeciesRaster.R
\name{createSpeciesRaster}
\alias{createSpeciesRaster}
\title{Create speciesRaster}
\usage{
createSpeciesRaster(ranges, rasterTemplate = NULL, verbose = FALSE)
}
\arguments{
\item{ranges}{Either a RasterStack, RasterBrick, or species by cell matrix. If 
raster objects, cell values can either be binary presence/absence, or probabilities.}

\item{rasterTemplate}{If input is a species x cell matrix, then a rasterTemplate must be
provided.}

\item{verbose}{Primarily intended for debugging, print progress to the console.}
}
\value{
an object of class \code{speciesRaster}.
}
\description{
This function takes a rasterStack and generates both a richness 
raster and an associated list of species per cell, creating an object of 
class \code{speciesRaster}.
}
\details{
This function generates an object of class \code{speciesRaster}, which is a 
list containing the following elements:
\itemize{
	\item{\code{raster:}} {A raster representing counts of species per cell.}
	\item{\code{speciesList:}} {A list of species found in each cell.}
	\item{\code{geogSpecies:}} {a vector of unique species in all cells.}
	\item{\code{data:}} {An empty spot that morphological data can be added to.}
	\item{\code{phylo:}} {An empty spot that a phylogeny can be added to.}
}

If input is a RasterStack, then all parameters are taken from that, such as resolution,
extent and projection. Any non-NA and non-zero cell is considered a presence.	
This function expects that all input rasters in the rasterStack have presence values
(i.e., at least 1 non-NA value). If any rasters have exclusively NA cells, then the 
function will stop with a warning, and the output will be the index in the rasterStack
of those rasters.
}
\examples{
library(raster)
library(maptools)
# example dataset: a list of 24 chipmunk distributions as polygons
head(tamiasPolyList)

# convert polygon ranges to raster
ranges <- rasterStackFromPolyList(tamiasPolyList, resolution = 20000)

spRas <- createSpeciesRaster(ranges = ranges)

spRas




}
\author{
Pascal Title
}
