% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.speciesRaster.R
\name{plot.speciesRaster}
\alias{plot.speciesRaster}
\title{Plot speciesRaster}
\usage{
\method{plot}{speciesRaster}(x, log = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{speciesRaster}}

\item{log}{boolean; should the cell values be logged?}

\item{...}{additional parameters will be passed to the raster package plot function.}
}
\value{
Nothing is returned.
}
\description{
Plot a speciesRaster object.
}
\examples{
plot(tamiasSpRas)

plot(tamiasSpRas, legend=FALSE, axes=FALSE, box=FALSE)
addRasterLegend(tamiasSpRas[[1]], location = 'top', ramp=c('blue','yellow','red'))

}
\author{
Pascal Title
}
