% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.speciesRaster.R
\name{plot.speciesRaster}
\alias{plot.speciesRaster}
\title{Plot speciesRaster}
\usage{
plot.speciesRaster(x, log = FALSE, includeLegend = TRUE, col = c("blue",
  "yellow", "red"), box = TRUE, axes = TRUE, location = "right", ...)
}
\arguments{
\item{x}{object of class \code{speciesRaster}}

\item{log}{boolean; should the cell values be logged?}

\item{includeLegend}{boolean; should legend be included?}

\item{col}{either a vector of color names that will be interpolated, or a color ramp
function that takes an integer (see for example \code{\link{colorRampPalette}})}

\item{box}{boolean; should box be drawn around plot?}

\item{axes}{boolean; should axes be included?}

\item{location}{location of legend, if included. See \code{\link{addRasterLegend}}.}

\item{...}{additional parameters will be passed to the \code{\link{addRasterLegend}} function.}
}
\value{
Nothing is returned.
}
\description{
Plot a speciesRaster object.
}
\examples{
plot(tamiasSpRas)

plot(tamiasSpRas, includeLegend=FALSE, axes=FALSE, box=FALSE)
addRasterLegend(tamiasSpRas, location = 'top', ramp=c('blue','yellow','red'))

}
\author{
Pascal Title
}
