% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterStackFromPolyList.R
\name{rasterStackFromPolyList}
\alias{rasterStackFromPolyList}
\title{Polygon List to rasterStack}
\usage{
rasterStackFromPolyList(polyList, resolution = 50000,
  retainSmallRanges = TRUE, extent = "auto", dropEmptyRasters = TRUE,
  nthreads = 1)
}
\arguments{
\item{polyList}{a list of SpatialPolygon objects, named with taxon names}

\item{resolution}{vertical and horizontal size of raster cell, in units 
of the polygons' projection}

\item{retainSmallRanges}{boolean; should small ranged species be dropped or preserved.
See details.}

\item{extent}{if 'auto', then the maximal extent of the polygons will be used. 
If not auto, must be a numeric vector of length 4 with minLong, maxLong, minLat, maxLat.}

\item{dropEmptyRasters}{if \code{TRUE}, then species that have no presence cells will be dropped.
If \code{FALSE}, then rasters will remain, filled entirely with \code{NA}.}

\item{nthreads}{number of threads to use for parallelization of the function.}
}
\value{
an object of class \code{RasterStack} where all rasters contain values of 
either NA or 1.
}
\description{
Takes a list of polygons and creates a rasterStack.
}
\details{
In the rasterization process, all cells for which the polygon covers the midpoint are
	considered as present and receive a value of 1. If \code{retainSmallRanges = FALSE}, 
	then species whose ranges are so small that no cell registers as present will be 
	dropped. If \code{retainSmallRanges = TRUE}, then the cells that the small polygon
	is found in will be considered as present.

	If \code{dropEmptyRasters = TRUE} and \code{retainSmallRanges = TRUE}, then the species that 
	will be dropped are those that are outside of the requested extent (which in that case 
	would be specified explicitly).
}
\examples{
library(raster)
library(maptools)
# example dataset: a list of 24 chipmunk distributions as polygons
head(tamiasPolyList)

rangeStack <- rasterStackFromPolyList(tamiasPolyList, resolution = 50000)
rangeStack

}
\author{
Pascal Title
}
